+++ CHANGELOG.md
@@ -79,6 +79,7 @@
 * nostr: remove `#[cfg]` attributes from `NostrSigner` NIP04 and NIP44 methods ([Yuki Kishimoto])
 * nostr: omit everything after the relay-hint in NIP22 `p` tags ([Yuki Kishimoto])
 * nostr: make EventBuilder `custom_created_at` and `pow` fields public ([Yuki Kishimoto])
+* nostr: improve `Filter::match_event` performance ([Yuki Kishimoto])
 * pool: update `Error::WebSocket` variant inner type ([Yuki Kishimoto])
 * pool: refactor negentropy sync methods ([Yuki Kishimoto])
 * pool: refactor relay pinger ([Yuki Kishimoto])